+++ MantidQt/CustomInterfaces/inc/MantidQtCustomInterfaces/DynamicPDF/SliceSelector.h
@@ -51,7 +51,7 @@ private slots:
   void launchBackgroundRemover();
 
 private:
-  void initLayout();
+  void initLayout() override;
   /// The form generated by Qt Designer
   Ui::SliceSelector m_uiForm;
   boost::shared_ptr<WorkspaceRecord> m_loadedWorkspace;
