+++ MantidQt/SliceViewer/inc/MantidQtSliceViewer/SliceViewer.ui
@@ -822,7 +822,7 @@
            <widget class="MantidQt::MantidWidgets::SafeQwtPlot" name="safeQwtPlot"/>
           </item>
           <item>
-           <widget class="MantidQt::SliceViewer::ColorBarWidget" name="colorBarWidget" native="true"/>
+           <widget class="MantidQt::MantidWidgets::ColorBarWidget" name="colorBarWidget" native="true"/>
           </item>
          </layout>
         </widget>
@@ -836,9 +836,9 @@
  <layoutdefault spacing="6" margin="11"/>
  <customwidgets>
   <customwidget>
-   <class>MantidQt::SliceViewer::ColorBarWidget</class>
+   <class>MantidQt::MantidWidgets::ColorBarWidget</class>
    <extends>QWidget</extends>
-   <header>MantidQtSliceViewer/ColorBarWidget.h</header>
+   <header>MantidQtMantidWidgets/ColorBarWidget.h</header>
   </customwidget>
   <customwidget>
    <class>QwtPlot</class>
